feature: Add Utils - Validator, Logger, Enum for Account Statuses, Transaction Types, User and UserAccount

+++ src/main/java/Laboratory_Work_1_SOLID_Principles/Interfaces/ITerminal.java
@@ -1,5 +1,5 @@
 package Laboratory_Work_1_SOLID_Principles.Interfaces;
 
 public interface ITerminal {
-    void performTransaction(IAccount account, ILogger logger, float amount);
+    void performTransaction(IAccount account, ILogger logger, double amount);
 }
